Added new methods to BaseHttpHandler to avoid code duplication in classes that extend it
