+++ src/httptaskserver/adapters/DurationAdapter.java
@@ -1,4 +1,4 @@
-package httptaskserver;
+package httptaskserver.adapters;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
