+++ src/httptaskserver/handlers/EpicHandler.java
@@ -1,14 +1,14 @@
-package httptaskserver;
+package httptaskserver.handlers;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.NotFoundException;
 import manager.TaskManager;
-import model.Epic;
 
 import java.io.IOException;
 
-class EpicHandler extends BaseHttpHandler implements HttpHandler {
+public class EpicHandler extends BaseHttpHandler implements HttpHandler {
+    final String taskType = "epic";
     public EpicHandler(TaskManager manager) {
         super(manager);
     }
@@ -19,32 +19,18 @@ class EpicHandler extends BaseHttpHandler implements HttpHandler {
         String uri = exchange.getRequestURI().getPath();
         String[] uriParts = uri.split("/");
         if (method.equals("GET") && uriParts.length == 2) {
-            String text = gson.toJson(manager.getEpics());
-            sendText(exchange, text, method);
+            getAllTasks(exchange, method, taskType);
         } else if (method.equals("GET") && uriParts.length == 3) {
             try {
-                int id = Integer.parseInt(uriParts[2]);
-                String text = gson.toJson(manager.getById(id));
-                sendText(exchange, text, method);
+                getCertainTask(exchange, method, uriParts[2]);
             } catch (NumberFormatException | NotFoundException e) {
                 sendNotFound(exchange);
             }
         } else if (method.equals("POST") && uriParts.length == 2) {
-            String input = new String(exchange.getRequestBody().readAllBytes(), defaultcharset);
-            Epic epic = gson.fromJson(input, Epic.class);
-            if (epic.getId() == 0) {
-                manager.addEpic(epic);
-                sendText(exchange, "Epic has been successfully added", "POST");
-            } else {
-                manager.updateEpic(epic);
-                sendText(exchange, "Epic has been successfully added", "POST");
-            }
+            postTask(exchange, taskType);
         } else if (method.equals("DELETE") && uriParts.length == 3) {
             try {
-                int id = Integer.parseInt(uriParts[2]);
-                manager.removeById(id);
-                String text = "The epic has been deleted";
-                sendText(exchange, text, method);
+                deleteTask(exchange, method, uriParts[2]);
             } catch (NumberFormatException | NotFoundException e) {
                 sendNotFound(exchange);
             }
