+++ tasks.csv
@@ -1,5 +1,9 @@
 id,type,name,status,description,duration,startTime,epic
+1,TASK,Задача 1,NEW,Детали задачи 1,100,2025.04.23 12:30
 2,TASK,Задача 1,NEW,Детали задачи 1,100,2025.04.23 14:30
 3,TASK,Задача 1,NEW,Детали задачи 1,100,2025.04.23 16:30
+6,TASK,Задача 1,NEW,Детали задачи 1,100,2025.04.24 12:30
 4,EPIC,Эпик A,DONE,Детали эпика A
+7,EPIC,Эпик A,NEW,Детали эпика A
 5,SUBTASK,Подзадача A1,DONE,Детали подзадачи A1,20,2025.04.23 20:00,4
+8,SUBTASK,Подзадача A1,DONE,Детали подзадачи A1,20,2025.04.24 16:00,4
