+++ test/HttpTaskManagerHistoryTest.java
@@ -1,6 +1,6 @@
-import httptaskserver.DurationAdapter;
+import httptaskserver.adapters.DurationAdapter;
 import httptaskserver.HttpTaskServer;
-import httptaskserver.LocalDateTimeAdapter;
+import httptaskserver.adapters.LocalDateTimeAdapter;
 import com.google.gson.*;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
